Update ALU.vhd
alu_A is the input of the ALU that comes from the Databus. This information may come directly from the instructioncode, from a register or from the output of the ALU.
alu_B is connected to the output of the ALU.
The opcode gives the ALU an instrunction on what calculation is should make. With every 4-bits value, a calculation is defined.